Migrate DurationValues
